security: fix 7 issues from audit (input validation, overflow, div-by-zero)
- cuda_pipeline: int32 → int64 for packet sizes (prevents overflow in sort kernel)
- congestion/red: validate min/max_thresh > 0, max_prob in (0,1], weight in (0,1);
  clamp gentle RED denominator to 1e-9 to prevent division by zero
- congestion/controller: cap burst size at 1MB (prevents unbounded token bucket)
- engine: validate load_factor, tcp/udp_loss_rate in [0,1], link_capacity > 0,
  raise ValueError on unknown scheduler (was silently defaulting to adaptive)
- tests: update weight=1.0 → weight=0.999 to comply with validated bounds

+++ slicenet/gpu/cuda_pipeline.py
@@ -86,7 +86,7 @@ class PacketBatch:
     def empty(cls, capacity: int) -> "PacketBatch":
         return cls(
             n=0,
-            sizes=np.zeros(capacity, dtype=np.int32),
+            sizes=np.zeros(capacity, dtype=np.int64),
             traffic_classes=np.zeros(capacity, dtype=np.int32),
             protocols=np.zeros(capacity, dtype=np.int8),
             priorities=np.zeros(capacity, dtype=np.int32),
